pyproject.toml
[project]
name = "aws-strands-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ddgs>=9.5.4",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "nest-asyncio>=1.6.0",
    "ollama>=0.5.3",
    "pydantic>=2.11.7",
    "strands-agents>=1.6.0",
    "strands-agents-tools>=0.2.5",
    "uvicorn>=0.35.0",
]
